refactor search and doc, split logic to providers
